move astro components to svelte
This commit moves all previous Astro components (not pages) to Svelte,
improving the authoring experience when using Astro islands.

A global Svelte store is used to track `Astro.url.pathname` because
direct access to `Astro.*` is not available inside Svelte components.

Alpine.js was also considered and enjoyed but the interoperability
between Astro's frontmatter and Alpine's string-interpreted data did not
prove to be too friendly.