add palette ingredients
This commit adds the palette ingredients. The ingredients can be copied
and formatted. A clipboard helper is included in the global store as
well as the ability to save the format preference. The format behaviour
is located in the new palette data.

Note, this commit uses a pre-release version of @rose-pine/palette.