add palette swatches and move ingredients to subpage
This commit moves ingredients to /palette/ingredients and adds the new
swatches to /palette. These pages share a layout to give the illusion of
tab navigation but without JavaScript. A specification with colour usage
is now exported from the palette data.

Slot animation can now be disabled in the page intro component to
account for the tab-like navigation.

Finally, new `.scheme-{role}` global classes have been added. These are
used for the specification tags.