add i18n
This commit introduces i18n through the astro-i18next integration. Some
translations are included but will be improved in future commits.

There are currently issues with JavaScript-enabled Svelte components and
translations. The only workaround needed at this time was to pass a
label property to the `ThemesFilter` component. If this grows into a
larger problem, we may need to explore another framework solution.