allow user page scaling
This commit removes `maximum-scale` as the accessibility consequences
were not known until reading through the recent lighthouse report. The
reason for adding this was to prevent automatic zooming for text inputs
with font sizes below 16px. To remedy this, our text inputs now use a
minimum font size of 16px.